Avoid generating TagLink ids ending with a colon
When writing edges to Neptune in lpg mode, if a "from" or a "to" vertex does not
exist, it's created. The label of those vertices is determined
by calling the function: "parse_arn" with the id of the vertex to be created.
The problem is the function returns empty is the value passed in has the
following shape: "value:". As empty labels are not supported
by Neptune that causes and error when writing the edge.