Fix writing the graph to Neptune as an lpg graph (#164)
* Use string type in edge ids instead of uuid

Neptune does not support the uuid type, so in order to ensure all
the ids of the edges are strings before writing them to Neptune
we directly use convert the uuid's to strings when setting the id's
of the edges when transforming the graph to Neptune.

* Avoid generating TagLink ids ending with a colon

When writing edges to Neptune in lpg mode, if a "from" or a "to" vertex does not
exist, it's created. The label of those vertices is determined
by calling the function: "parse_arn" with the id of the vertex to be created.
The problem is the function returns empty is the value passed in has the
following shape: "value:". As empty labels are not supported
by Neptune that causes and error when writing the edge.

* Ensure the arn's written to Neptune are not of the type uuid